Split shared scanner helpers by responsibility
