+++ src/core/calc/scanner_nk.cpp
@@ -7,7 +7,11 @@
 #include "core/calc/scanner_nk.h"
 
 #include "core/calc/cosine_distance.h"
-#include "core/calc/scanner_engine_common.h"
+#include "core/calc/scanner_dataset_scan.h"
+#include "core/calc/scanner_heap_utils.h"
+#include "core/calc/scanner_log_utils.h"
+#include "core/calc/scanner_query_context.h"
+#include "core/utils/timer.h"
 
 #include "numkong/capabilities.h"
 #include "numkong/reduce.h"
